requirements.txt
Flask
flask-cors
tensorflow
pillow
numpy
gunicorn
